+++ lib/sendgrid4r.rb
@@ -108,14 +108,14 @@ require 'sendgrid4r/cli/stats/subuser'
 require 'sendgrid4r/cli/stats/stats'
 require 'sendgrid4r/cli/subusers/monitor'
 require 'sendgrid4r/cli/subusers/subuser'
-require 'sendgrid4r/cli/supressions/block'
-require 'sendgrid4r/cli/supressions/bounce'
-require 'sendgrid4r/cli/supressions/global_unsubscribe'
-require 'sendgrid4r/cli/supressions/group_unsubscribe'
-require 'sendgrid4r/cli/supressions/group'
-require 'sendgrid4r/cli/supressions/invalid_email'
-require 'sendgrid4r/cli/supressions/spam_report'
-require 'sendgrid4r/cli/supressions/supression'
+require 'sendgrid4r/cli/suppressions/block'
+require 'sendgrid4r/cli/suppressions/bounce'
+require 'sendgrid4r/cli/suppressions/global_unsubscribe'
+require 'sendgrid4r/cli/suppressions/group_unsubscribe'
+require 'sendgrid4r/cli/suppressions/group'
+require 'sendgrid4r/cli/suppressions/invalid_email'
+require 'sendgrid4r/cli/suppressions/spam_report'
+require 'sendgrid4r/cli/suppressions/suppression'
 require 'sendgrid4r/cli/templates/version'
 require 'sendgrid4r/cli/templates/template'
 require 'sendgrid4r/cli/webhooks/event'
