+++ lib/sendgrid4r/cli/suppressions/spam_report.rb
@@ -1,14 +1,14 @@
 module SendGrid4r::CLI
-  module Supressions
-    class InvalidEmail < SgThor
+  module Suppressions
+    class SpamReport < SgThor
 
-      desc 'list', 'List blocks'
+      desc 'list', 'List spam reports'
       option :start_time, :type => :numeric
       option :end_time, :type => :numeric
       option :limit, :type => :numeric
       option :offset, :type => :numeric
       def list
-        puts @client.get_invalid_emails(
+        puts @client.get_spam_reports(
           start_time: options[:start_time], end_time: options[:end_time],
           limit: options[:limit], offset: options[:offset]
         )
@@ -16,15 +16,15 @@ module SendGrid4r::CLI
         puts e.inspect
       end
 
-      desc 'delete', 'Delete invalid emails'
+      desc 'delete', 'Delete spam reports'
       option :delete_all, :type => :boolean
       option :email
       option :emails, :type => :array
       def delete
         if options[:email]
-          @client.delete_invalid_email(email: options[:email])
+          @client.delete_spam_report(email: options[:email])
         else
-          @client.delete_invalid_emails(
+          @client.delete_spam_reports(
             delete_all: options[:delete_all], emails: options[:emails]
           )
         end
@@ -32,10 +32,10 @@ module SendGrid4r::CLI
         puts e.inspect
       end
 
-      desc 'get', 'Get a invalid email'
+      desc 'get', 'Get a spam report'
       option :email, :require => true
       def get
-        puts @client.get_invalid_email(email: options[:email])
+        puts @client.get_spam_report(email: options[:email])
       rescue RestClient::ExceptionWithResponse => e
         puts e.inspect
       end
