+++ spec/cli/suppressions/block_spec.rb
@@ -1,8 +1,8 @@
 # encoding: utf-8
 require File.dirname(__FILE__) + '/../../spec_helper'
 
-module SendGrid4r::CLI::Supressions
-  describe Bounce do
+module SendGrid4r::CLI::Suppressions
+  describe Block do
     describe 'integration test', :it do
       before do
         Dotenv.load
@@ -13,9 +13,11 @@ module SendGrid4r::CLI::Supressions
           'list',
           '--apikey', ENV['SILVER_API_KEY'],
           '--start_time', Time.local(2016, 1, 1).to_i,
-          '--end_time', Time.local(2016, 1, 31).to_i
+          '--end_time', Time.local(2016, 1, 31).to_i,
+          '--limit', 10,
+          '--offset', 0
         ]
-        Bounce.start(args)
+        Block.start(args)
       end
 
       it '#delete with delete_all' do
@@ -24,7 +26,7 @@ module SendGrid4r::CLI::Supressions
           '--apikey', ENV['SILVER_API_KEY'],
           '--delete_all', true
         ]
-        Bounce.start(args)
+        Block.start(args)
       end
 
       it '#delete with email' do
@@ -33,7 +35,7 @@ module SendGrid4r::CLI::Supressions
           '--apikey', ENV['SILVER_API_KEY'],
           '--email', ''
         ]
-        Bounce.start(args)
+        Block.start(args)
       end
 
       it '#delete with emails' do
@@ -42,7 +44,7 @@ module SendGrid4r::CLI::Supressions
           '--apikey', ENV['SILVER_API_KEY'],
           '--emails', 'abc@abc.com', 'cde@cde.com'
         ]
-        Bounce.start(args)
+        Block.start(args)
       end
 
       it '#get' do
@@ -51,7 +53,7 @@ module SendGrid4r::CLI::Supressions
           '--apikey', ENV['SILVER_API_KEY'],
           '--email', ''
         ]
-        Bounce.start(args)
+        Block.start(args)
       end
     end
   end
