+++ spec/cli/suppressions/bounce_spec.rb
@@ -1,8 +1,8 @@
 # encoding: utf-8
 require File.dirname(__FILE__) + '/../../spec_helper'
 
-module SendGrid4r::CLI::Supressions
-  describe InvalidEmail do
+module SendGrid4r::CLI::Suppressions
+  describe Bounce do
     describe 'integration test', :it do
       before do
         Dotenv.load
@@ -13,11 +13,9 @@ module SendGrid4r::CLI::Supressions
           'list',
           '--apikey', ENV['SILVER_API_KEY'],
           '--start_time', Time.local(2016, 1, 1).to_i,
-          '--end_time', Time.local(2016, 1, 31).to_i,
-          '--limit', 10,
-          '--offset', 0
+          '--end_time', Time.local(2016, 1, 31).to_i
         ]
-        InvalidEmail.start(args)
+        Bounce.start(args)
       end
 
       it '#delete with delete_all' do
@@ -26,16 +24,16 @@ module SendGrid4r::CLI::Supressions
           '--apikey', ENV['SILVER_API_KEY'],
           '--delete_all', true
         ]
-        InvalidEmail.start(args)
+        Bounce.start(args)
       end
 
       it '#delete with email' do
         args = [
           'delete',
           '--apikey', ENV['SILVER_API_KEY'],
-          '--email', 'abc@abc.com'
+          '--email', ''
         ]
-        InvalidEmail.start(args)
+        Bounce.start(args)
       end
 
       it '#delete with emails' do
@@ -44,16 +42,16 @@ module SendGrid4r::CLI::Supressions
           '--apikey', ENV['SILVER_API_KEY'],
           '--emails', 'abc@abc.com', 'cde@cde.com'
         ]
-        InvalidEmail.start(args)
+        Bounce.start(args)
       end
 
       it '#get' do
         args = [
           'get',
           '--apikey', ENV['SILVER_API_KEY'],
-          '--email', 'abc@abc.com'
+          '--email', ''
         ]
-        InvalidEmail.start(args)
+        Bounce.start(args)
       end
     end
   end
