+++ spec/cli/suppressions/group_unsubscribe_spec.rb
@@ -1,7 +1,7 @@
 # encoding: utf-8
 require File.dirname(__FILE__) + '/../../spec_helper'
 
-module SendGrid4r::CLI::Supressions
+module SendGrid4r::CLI::Suppressions
   describe GroupUnsubscribe do
     describe 'integration test', :it do
       before do
